refact: 상세페이지 파라미터 data class로 분리, 파싱로직 추가/ goods하위 변수명,타입 변경

+++ lib/src/features/order/presentation/widget/order_info_item.dart
@@ -19,8 +19,8 @@ class OrderInfoItem extends StatelessWidget {
   final String imageUrl;
   final String storeName;
   final String address;
-  final double discount;
-  final double price;
+  final int discount;
+  final int price;
   final int count;
   final double imageSize;
